layout without new widget class
