+++ Sliders.py
@@ -6,7 +6,7 @@ import holoviews as hv
 
 import sys
  
-from PyQt5.QtWidgets import QApplication, QMainWindow, QMenu, QVBoxLayout, QSizePolicy, QMessageBox, QWidget, QPushButton, QSlider, QHBoxLayout
+from PyQt5.QtWidgets import QApplication, QMainWindow, QMenu, QVBoxLayout, QSizePolicy, QMessageBox, QWidget, QPushButton, QSlider, QHBoxLayout, QGridLayout
 from PyQt5.QtGui import QIcon
 from PyQt5.QtCore import Qt,  QCoreApplication
  
@@ -16,8 +16,6 @@ import time
 import numpy as np
 
 import matplotlib.pyplot as plt
-
-from SliderGroup import SliderGroup
 
 from matplotlib.backends.qt_compat import QtCore, QtWidgets, is_pyqt5
 if is_pyqt5():
@@ -57,13 +55,15 @@ class ApplicationWindow(QtWidgets.QMainWindow):
         # Create a layout
         self._main = QtWidgets.QWidget()
         self.setCentralWidget(self._main)
-        layout = QtWidgets.QVBoxLayout(self._main)
+        # layout = QtWidgets.QVBoxLayout(self._main)
 
         # Create a FigureCanvas
         self.canvas = FigureCanvas(self.figure)
-        
+       
+        layout = QGridLayout(self._main)
+ 
         # Add the canvas to the layout
-        layout.addWidget(self.canvas)
+        layout.addWidget(self.canvas,0,0,1,3)
         self.addToolBar(NavigationToolbar(self.canvas, self))
         
         # Create a list of horizontal sliders
@@ -71,8 +71,56 @@ class ApplicationWindow(QtWidgets.QMainWindow):
         
         for i in range(n_dims):
         
-            layout.addWidget(SliderGroup(i,self.arr,self.ax,self.canvas))
+            # Create a slider for dimension-i
+            self.sliders[i] = self.create_slider(i)
+
+            # Create X and Y buttons for the slider
+            x_button, y_button = self.create_buttons(i)
+
+            # Add the buttons and slider to the box
+            layout.addWidget(x_button, i+1, 0)
+            layout.addWidget(y_button, i+1, 1)
+            layout.addWidget(self.sliders[i], i+1, 2)
+        
+    def create_slider(self, n_dim):
+    
+        sl = QSlider(Qt.Horizontal)
+        sl.valueChanged.connect(self.value_change(n_dim))
+      
+        # Set slider values
+        sl.setMinimum(0)
+        sl.setMaximum(self.arr.shape[n_dim] - 1)
+        sl.setValue(0)
+        
+        # Set tick interval
+        sl.setTickInterval(1)
             
+        # Disable slider by default
+        # sl.setEnabled(False)
+        
+        return sl
+    
+    def create_buttons(self, n_dim):
+   
+        # todo - make the buttons do something 
+        return [QPushButton("X"), QPushButton("Y")]
+    
+    def value_change(self,n_dim):
+   
+        # totally clear names 
+        def value_changer():
+
+            # Obtain the slider value
+            slider_val = self.slider.value()
+      
+            # Change plot
+            self.ax.imshow(self.arr.take(indices=slider_val,axis=n_dim))
+           
+            # Will this work? 
+            self.canvas.draw()
+        
+        return value_changer
+        
 if __name__ == "__main__":
     qapp = QtWidgets.QApplication(sys.argv)
     app = ApplicationWindow()
